mejoras varias
principal mejora, las armas se detectan como los hijos de los blobs de naves de primer nivel
se arreglo un bug donde la rotacion del escudo no afectaba al collider (porque sucedia despues de agregar el collider)
se agrego la opcion de ignorar blobs al configurar templates
se arreglo que en la partida no se cambien los script objects default sino que genera instancias

+++ Assets/UI/HSVFilterPreview.shader
@@ -92,7 +92,7 @@ Shader "Unlit/HSVFilterPreview"
                 hsv.g = GammaToLinearSpace(hsv.g);
                 hsv.b = GammaToLinearSpace(hsv.b);
                 
-                fixed3 enRango = step(_HSVMin, hsv.rgb);
+                float3 enRango = step(_HSVMin, hsv.rgb);
 
                 enRango = enRango * step(hsv.rgb, _HSVMax);
                 enRango = lerp(enRango, 1.0-enRango, _Invertir);
